requirements.txt
aiodns==3.0.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
backports.zoneinfo==0.2.1
ccxt==1.56.29
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
cryptography==3.4.8
cycler==0.10.0
dateparser==1.0.0
Flask==2.0.1
idna==3.2
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
multidict==5.1.0
numpy==1.21.2
parinya==0.0.27
Pillow==8.3.2
pycares==4.0.0
pycparser==2.20
pyparsing==2.4.7
python-binance==1.0.14
python-dateutil==2.8.2
python-decouple==3.4
python-dotenv==0.19.0
pytz==2021.1
regex==2021.8.28
requests==2.26.0
six==1.16.0
typing-extensions==3.10.0.2
tzlocal==3.0
ujson==4.1.0
urllib3==1.26.6
websockets==10.0
Werkzeug==2.0.1
yarl==1.6.3
zipp==3.5.0
